Detect and display mod key based on user agent
On mac, the modifier key is `cmd` while other systems will most likely use
`ctrl`. The underlying keybinding plugin handles this situation already
but Octo did not adjust the visual hints in the UI for non mac users.

To solve this, we look for `Macintosh` in the user agent. If present,
display `cmd` as the mod key in keybinding hints. Otherwise, we assume
`ctrl`.

This is calculated when initializing the app. We set this state in the
store.