Add 'Daily Notepad' feature
This page will always return the current daily doc or create a new one if it does not exist. The cutoff time is 3:00 AM (client time) to account for night-owls (like myself). After 3am, the daily doc page will create a new doc for the next day. This should probably be configurable.